// projects/addons/var/error/interfaces.ts
/*
 * Copyright (c) 2016-2026 Broadcom. All Rights Reserved.
 * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
 * This software is released under MIT license.
 * The full license information can be found in LICENSE in the root directory of this project.
 */

export interface ActivationError {
  data: ErrorData;
}
export interface ErrorData {
  message?: string;
  stackTrace?: string;
}